% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jewels-package.R
\docType{package}
\name{gemstones-package}
\alias{gemstones}
\alias{gemstones-package}
\title{gemstones: Make Your 'golem' App Shine}
\description{
Make Your 'golem' app shine.
}
\author{
\strong{Maintainer}: Colin Fay \email{contact@colinfay.me} (\href{https://orcid.org/0000-0001-7343-1846}{ORCID})

Other contributors:
\itemize{
  \item ThinkR [copyright holder]
}

}
\keyword{internal}
