% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_jqueryi18next.R
\name{use_jqueryi18next}
\alias{use_jqueryi18next}
\title{use_jqueryi18next Title}
\usage{
use_jqueryi18next(
  pkg = golem::get_golem_wd(),
  i18next_version = "21.6.11",
  jquery_i18next_version = "1.2.1",
  with_test = TRUE
)
}
\value{
1
}
\description{
use_jqueryi18next Title
}
\examples{
if (interactive()) {
  use_jqueryi18next()
}
}
