% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_notifyjs.R
\name{use_notifyjs}
\alias{use_notifyjs}
\title{Add notifyjs}
\usage{
use_notifyjs(pkg = golem::get_golem_wd(), with_test = TRUE)
}
\arguments{
\item{pkg}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{with_test}{should the module be created with tests?}
}
\value{
TRUE invisibly
}
\description{
Add notifyjs to your app
}
\examples{
if (interactive()) {
  use_notifyjs()
}
}
