% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_jqueryi18next.R
\name{use_jqueryi18next}
\alias{use_jqueryi18next}
\title{use_jqueryi18next}
\usage{
use_jqueryi18next(
  pkg = golem::get_golem_wd(),
  i18next_version = "21.6.11",
  jquery_i18next_version = "1.2.1",
  with_test = TRUE
)
}
\arguments{
\item{pkg}{Path to the root of the package. Default is \code{get_golem_wd()}.}

\item{i18next_version, jquery_i18next_version}{Version of the modules}

\item{with_test}{should the module be created with tests?}
}
\value{
TRUE invisibly
}
\description{
Add jqueryi18next to your app
}
\examples{
if (interactive()) {
  use_jqueryi18next()
}
}
